pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bind9-ctl"
version = "0.1.0"
description = "Declarative DNS zone management CLI for BIND9."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.6",
    "Jinja2>=3.1",
    "PyYAML>=6.0",
    "pydantic>=2.6",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
]

[project.scripts]
bind9-ctl = "bind9_ctl.cli:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",
    "F",
    "I",
    "UP",
    "B",
    "C4",
    "DTZ",
    "PIE",
    "PT",
    "PTH",
    "RUF",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

